Add files via upload
This is a copy of the APK Manifest file for the Android app.  As I said in the README.md file, this Android app was created via an online service, so my next step would be to create one myself on my own PC using Linux tools so I have tighter control of ads and permissions.  For this manifest I used jadx/jadx gui to reverse engineer the obfuscated apk file.